// src/api/generated/model/apiResponseListBrandDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { BrandDto } from "./brandDto";

export interface ApiResponseListBrandDto {
  success?: boolean;
  message?: string;
  data?: BrandDto[];
  errors?: string[];
  timestamp?: string;
  path?: string;
}