// src/api/generated/model/apiResponseUserDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { UserDto } from "./userDto";

export interface ApiResponseUserDto {
  success?: boolean;
  message?: string;
  data?: UserDto;
  errors?: string[];
  timestamp?: string;
  path?: string;
}